+++ test/unit/change-event-buffer.test.ts
@@ -38,7 +38,7 @@ config.parallel('change-event-buffer.test.js', () => {
             const last = schemaObjects.human();
             await col.insert(last);
             const lastBufferEvent = col._changeEventBuffer.buffer[col._changeEventBuffer.buffer.length - 1];
-            assert.strictEqual(last.passportId, lastBufferEvent.data.v.passportId);
+            assert.strictEqual(last.passportId, lastBufferEvent.documentData.passportId);
 
             col.database.destroy();
         });
@@ -101,7 +101,7 @@ config.parallel('change-event-buffer.test.js', () => {
             await col.insert(lastDoc);
 
             const gotIndex: any = col._changeEventBuffer.getArrayIndexByPointer(col._changeEventBuffer.counter);
-            assert.strictEqual(col._changeEventBuffer.buffer[gotIndex].data.v.firstName, lastDoc.firstName);
+            assert.strictEqual(col._changeEventBuffer.buffer[gotIndex].documentData.firstName, lastDoc.firstName);
 
             col.database.destroy();
         });
@@ -171,7 +171,7 @@ config.parallel('change-event-buffer.test.js', () => {
 
             const evs: any[] = col._changeEventBuffer.getFrom(q._latestChangeEvent + 1) as any;
             assert.strictEqual(evs.length, 1);
-            assert.strictEqual(evs[0].data.op, 'REMOVE');
+            assert.strictEqual(evs[0].operation, 'DELETE');
 
             col.database.destroy();
         });
@@ -192,8 +192,8 @@ config.parallel('change-event-buffer.test.js', () => {
             const reduced = q.collection._changeEventBuffer.reduceByLastOfDoc(allEvents);
 
             assert.strictEqual(reduced.length, 5);
-            const lastEvent: any = reduced.find(cE => cE.data.doc === oneDoc.primary);
-            assert.strictEqual(lastEvent.data.v.age, 5);
+            const lastEvent: any = reduced.find(cE => cE.documentId === oneDoc.primary);
+            assert.strictEqual(lastEvent.documentData.age, 5);
             col.database.destroy();
         });
     });